further progress, still buggy
Some async improvements.
Attempts to add a HD button.
Removed one superfluous view layer.
Attempt to insert a key loop (we should discuss what would how exactly this is supposed to work and whether it can work at all in WebKit - that remains unclear to me).
Autoplay support.
Remembers volume level (suggested on the bug tracker).

nagging issues:
The view doesn't seem to be redrawn when changing tabs.
I'd like to resize the view when knowing the movie size as the buttons may end up strangely positioned otherwise, so far I could only stall Safari when doing so (perhaps because I'm not on the main thread?)