Improve handling of embedded YouTube.
We were scanning for »var swfArgs = {« but that seems to fail on quite a few pages I saw last week. (e.g.:
http://www.fscklog.com/2009/10/drei-neue-get-a-mac-ads-zum-start-von-windows-7-videos.html 
http://www.scienceblogs.de/mathlog/2009/10/mathematik-zum-anfassen-auf-youtube.php 
http://www.fakesteve.net/2009/10/mosspuppet-claims-hes-sabotaged.html

It looks like the flashvars are stored after a »'SWF_ARGS': {« for those, so I'm scanning for that as well now.

A general YouTube change, perhaps? Does the old string still occur?