Open film asynchronously.
This used to be synchronous with the idea to display a progress indicator while things are loading. But it seems that synchronous initialising of the QTMovie likes giving nil results when things are going slowly.

Removing that bit of potential of clever here seems to make things work better. Even though there can still be a moment's blank if the downloading the film goes slowly.