Add menu item to toggle CtF to the Menu bar menu.
Does this work on X.4? (I have a hunch that bindings weren't able to do this back then)

Adjusts CtF's German terminology to match that of Safari's popup blocker. 

Should we provide a keyboard equivalent (say: Cmd-Opt-K or Cmd-Ctrl-K) for this out of the box or leave it to people to set it up for themselves?

Probably addresses this issue:

http://rentzsch.lighthouseapp.com/projects/24342/tickets/408-move-enable-clicktoflash#ticket-408-3