Ignore mouse clicks while in the background.
Both for the main view and our little buttons.

Reacting to background clicks is one of those bad OS X trends. And as CtF views can be quite large, it's particularly painful here, so we should do this properly.

(In case I have overlooked a way of doing this without subclassing NSButton, let me know!)