Let QTMovieView’s subviews want a layer.
QTMovieView seems to have two subviews, one for the film, the other for the controller. The former wants a layer the latter doesn't (on 10.6/64bit) and it tends to become invisible after switching tabs or scrolling in Safari without that.

Setting up all subviews to want a layer resolves the practical problem.