Add a view containing the buttons
This simplifies repositioning when our view is resized. It also lets us move the buttons around inside the plugin in one go. I want to do that for films, to make sure the buttons aren't drawn half on the film and half out of it.

Also do some cleanup along the way.