Attempt to hide the menu bar mor rigorously.
It seems preferable to hide the menu bar as it doesn't contain useful commands when in full-screen mode and it's easily invoked accidentally by hitting the top of the screen with the mouse and then covers the button for ending fullscreen mode.

Strangely this change doesn't actually improve the behaviour (on 10.6.2).