Improve autoplay logic.
We do autoplay ourselves as it seemed to fail for me occasionally with the native QT setting (no idea why). Interestingly this way of doing autoplay can be found in Apple example code, so this may be not as absurd as it seems.

The new addition is that we now ensure that we only autoplay once. This is important as in certain slow-loading situations the movie's state can toggle between the Playable and PlaythroughOK states. If not enough data of the film has been loaded, playback will stop, when more data comes in and the film switches back to the PlathroughOK state, playback would start again, resulting in a rather unnerving play/pause pattern which makes it impossible to let the film load in the background quietly.