Add ability to reload a video’s hash
To support this new feature in CTFKillerVideo which lets us deals with expired hash values.

Also make setting of the flashVars synchronous inside -_retrieveEmbeddedPlayerFlashVarsAndCheckForVariantsWithVideoId: so the newly retrieved values are reliably available when the method returns. 

To make that work properly, make CTFLoader run its NSURLConnections on the main thread rather than the current one (which may have stopped existing by the time the delegate methods need to be called).