More consistent shades of white.
As noted in 

http://rentzsch.lighthouseapp.com/projects/24342/tickets/492-the-white-colour-around-the-gear-icon-on-black-backgrounds-should-be-darker 

we don't use a consistent shade/opacity of white throughout the plug-in. Using a central method to provide that number, adapted to the current circumstances (preview image, active) now.

Going for less transparent drawing as (a) it seems that it was intended like that to begin with (the badge was drawn with alpha^2 opacity a) and (b) as illustrated in #492’s screenshots the previous version was hard to see on a dark background.