Handle save errors with vimeo videos.
It seems that QT stops refusing to save them when we remove the first frame. So we do that. It's not ideal but with the ‘artsy’ folks at vimeo chances seem good that the beginning of the film is solid black for a few frames anyway.

Of course finding a proper way to do this would still be preferable. But I don't see how that'd be possible in QT.

Addresses: http://rentzsch.lighthouseapp.com/projects/24342-clicktoflash/tickets