Clean up CTFKillerSIFR.
- now converts using the standard -convert method
- supports 'immediate conversion' (for the deSIFR case)

This solves the additional sIFR issues I reported in #507.

It does not solve #506 (closer investigation showed that #506 is in part caused by the site: their sIFR.rollback() function destroys the page even without CtF, so the only way around that would be not to call it. Ideas on how to determine when it's unsafe to call that function are appreciated.)